Backend/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
